# samples/sequelize/src/routes/auth.routes.ts
import { Router } from "express";
import bcrypt from "bcryptjs";
import jwt, { type SignOptions } from "jsonwebtoken";
import { registerDefinition } from "swaggiffy";
import { User } from "../models/User";

const router = Router();

const JWT_SECRET = process.env.JWT_SECRET || "secret";
const JWT_EXPIRES_IN = (process.env.JWT_EXPIRES_IN || "24h") as SignOptions["expiresIn"];

router.post("/register", async (req, res) => {
    try {
        const { name, email, password } = req.body;
        const exists = await User.findOne({ where: { email } });
        if (exists) return res.status(409).json({ error: "Email already registered" });

        const hashed = await bcrypt.hash(password, 10);
        const user = await User.create({ name, email, password: hashed });
        const token = jwt.sign(
            { id: user.id, email: user.email },
            JWT_SECRET,
            { expiresIn: JWT_EXPIRES_IN },
        );
        return res.status(201).json({ token, user: { id: user.id, name: user.name, email: user.email } });
    } catch (err: any) {
        return res.status(500).json({ error: err.message });
    }
});

router.post("/login", async (req, res) => {
    try {
        const { email, password } = req.body;
        const user = await User.findOne({ where: { email } });
        if (!user) return res.status(401).json({ error: "Invalid credentials" });

        const valid = await bcrypt.compare(password, user.password);
        if (!valid) return res.status(401).json({ error: "Invalid credentials" });

        const token = jwt.sign(
            { id: user.id, email: user.email },
            JWT_SECRET,
            { expiresIn: JWT_EXPIRES_IN },
        );
        return res.json({ token, user: { id: user.id, name: user.name, email: user.email } });
    } catch (err: any) {
        return res.status(500).json({ error: err.message });
    }
});

// Public routes — no security field on generated operations
registerDefinition(router, {
    basePath: "/api/auth",
    mappedSchema: "User",
    tags: "Auth",
    summary: "Authentication endpoints (public)",
    description: "Register and login do not require a bearer token.",
});

export { router as authRouter };